+++ mitra_cli/src/commands/mod.rs
@@ -1,3 +1,4 @@
 pub mod account;
 pub mod activitypub;
 pub mod filter;
+pub mod storage;